Completed trivia command and made small changes to other commands

+++ src/commands/Moderation/purge.js
@@ -30,6 +30,15 @@ module.exports = {
         const messages = interaction.options.getInteger("number-of-messages");
         const oldOnly = interaction.options.getBoolean("only-old") || false;
 
+        if(messages >= 100) return await interaction.editReply({
+            embeds: [
+                new EmbedBuilder()
+                    .setTitle("Invalid Input")
+                    .setDescription("The Messages cannot be more than or equal to 100")
+                    .setColor(0xdf2c14)
+            ]
+        })
+
         await interaction.channel.bulkDelete(messages, [ oldOnly ]);
 
         await interaction.channel.send({
